Implement candidate-generator lighters.
A new property `lighter-function` was added to the
p-search-candidate-generator type.  If provided, it is called in order
to retrieve a string to display as the buffer name.

It currently only uses the first candidate generator for the name. If
all candidate generators were included, the buffername could quickly
become very long.  I don't know if very long buffer names are okay.

All current candidate generators were also updated with a new lighter-function.